feat(chat): add ContextMentionChips and QueuedMessagePanel components
