+++ src/routes/Chat/ChatComposer.tsx
@@ -5,14 +5,16 @@ import type { PromptInputMessage } from "@/components/ai-elements/prompt-input"
 import type { QueuedChatMessage } from "@/components/app-shell/chat-queue"
 import type { ChatStatus } from "ai"
 
-import { AlertTriangle, ChevronRight, File as FileIcon, Folder, ListChecks, Package, Plug, Plus, X } from "lucide-react"
+import { AlertTriangle, File as FileIcon, Folder, Plus } from "lucide-react"
 import * as React from "react"
 import { AttachmentList } from "./ChatAttachments.tsx"
 import { buildConnectionPaletteItems, buildSkillPaletteItems, slashCommandItems } from "./composer-palette-items.ts"
-import { composerReducer, contextMentionKey, initialComposerState } from "./composer-state.ts"
+import { composerReducer, initialComposerState } from "./composer-state.ts"
 import { ComposerPalette } from "./ComposerPalette.tsx"
 import { ComposerTrailingControls } from "./ComposerTrailingControls.tsx"
+import { ContextMentionChips } from "./ContextMentionChips.tsx"
 import { AddCustomModelDialog } from "./ModelControls.tsx"
+import { QueuedMessagePanel } from "./QueuedMessagePanel.tsx"
 import { stripDraftAttachment, useComposerAttachments } from "./useComposerAttachments.ts"
 import { useComposerPalette } from "./useComposerPalette.ts"
 import { useModelCatalog } from "./useModelCatalog.ts"
@@ -27,7 +29,6 @@ import {
 } from "@/components/ai-elements/prompt-input"
 import { useSkillInventoryResource } from "@/components/AppDataHooks"
 import { Button } from "@/components/ui/button"
-import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@/components/ui/collapsible"
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu"
 import { useT } from "@/i18n/i18n"
 import { cn } from "@/lib/utils"
@@ -50,130 +51,6 @@ interface ChatComposerProps {
   ) => Promise<boolean>
   onStop: () => void
   onViewBilling?: () => void
-}
-
-function contextMentionLabel(mention: ChatContextMention): string {
-  return mention.kind === "skill" ? mention.name : mention.displayName
-}
-
-function ContextMentionChips({
-  mentions,
-  onRemove,
-}: {
-  mentions: ChatContextMention[]
-  onRemove?: (mention: ChatContextMention) => void
-}) {
-  const t = useT()
-  if (mentions.length === 0) {
-    return null
-  }
-  return (
-    <div className="flex w-full flex-wrap gap-2">
-      {mentions.map((mention) => (
-        <span
-          key={contextMentionKey(mention)}
-          className="oo-border-divider flex h-8 max-w-full items-center gap-2 rounded-lg border bg-background/70 px-2 text-sm shadow-xs"
-          title={mention.kind === "skill" ? mention.description : mention.accountLabel}
-        >
-          <span className="flex size-5 shrink-0 items-center justify-center rounded-md bg-muted text-muted-foreground">
-            {mention.kind === "skill" ? <Package className="size-3.5" /> : <Plug className="size-3.5" />}
-          </span>
-          <span className="min-w-0 truncate">
-            <span className="text-muted-foreground">
-              {mention.kind === "skill" ? t("chat.contextSkillPrefix") : t("chat.contextConnectionPrefix")}
-            </span>
-            <span className="ml-1 font-medium text-foreground">{contextMentionLabel(mention)}</span>
-          </span>
-          {onRemove ? (
-            <button
-              type="button"
-              aria-label={t("chat.contextRemove", { name: contextMentionLabel(mention) })}
-              className="-mr-1 flex size-5 shrink-0 items-center justify-center rounded-full text-muted-foreground hover:bg-muted hover:text-foreground"
-              onClick={() => onRemove(mention)}
-            >
-              <X className="size-3.5" />
-            </button>
-          ) : null}
-        </span>
-      ))}
-    </div>
-  )
-}
-
-function queuedMessagePreview(message: QueuedChatMessage): string {
-  const text = message.text.trim()
-  if (text) {
-    return text
-  }
-  return message.attachments.map((attachment) => attachment.name).join(", ")
-}
-
-function QueuedMessagePanel({ messages, onRemove }: { messages: QueuedChatMessage[]; onRemove: (id: string) => void }) {
-  const t = useT()
-  const [open, setOpen] = React.useState(true)
-  if (messages.length === 0) {
-    return null
-  }
-
-  return (
-    <Collapsible
-      open={open}
-      onOpenChange={setOpen}
-      className="oo-border-divider overflow-hidden rounded-xl border bg-background/95 shadow-xs backdrop-blur supports-[backdrop-filter]:bg-background/85"
-    >
-      <div className={cn("flex h-9 items-center px-2", open && "border-b border-border/50")}>
-        <CollapsibleTrigger asChild>
-          <button
-            type="button"
-            className="flex min-w-0 flex-1 items-center gap-2 rounded-lg px-1.5 py-1 text-left hover:bg-accent/45 focus-visible:border-ring focus-visible:ring-[3px] focus-visible:ring-ring/50 focus-visible:outline-none"
-            aria-label={open ? t("chat.queueCollapse") : t("chat.queueExpand")}
-          >
-            <ListChecks className="size-4 shrink-0 text-muted-foreground" />
-            <span className="oo-text-control min-w-0 flex-1 truncate text-muted-foreground">
-              {t("chat.queueTitle", { count: messages.length })}
-            </span>
-            <ChevronRight
-              className={cn("size-3.5 shrink-0 text-muted-foreground transition-transform", open && "rotate-90")}
-            />
-          </button>
-        </CollapsibleTrigger>
-      </div>
-      <CollapsibleContent>
-        <div className="max-h-40 overflow-auto">
-          {messages.map((message) => {
-            const preview = queuedMessagePreview(message)
-            return (
-              <div key={message.id} className="flex h-10 items-center gap-2 px-3">
-                <div className="min-w-0 flex-1">
-                  <div className="flex min-w-0 items-center gap-2">
-                    <span className="oo-text-control min-w-0 truncate text-foreground/90">
-                      {preview || t("chat.queueAttachmentOnly")}
-                    </span>
-                  </div>
-                  {message.attachments.length > 0 ? (
-                    <div className="oo-text-caption mt-0.5 truncate text-muted-foreground">
-                      {t("chat.queueAttachments", { count: message.attachments.length })}
-                    </div>
-                  ) : null}
-                </div>
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="icon"
-                  className="size-7 shrink-0 rounded-full text-muted-foreground hover:text-foreground"
-                  title={t("chat.queueRemove")}
-                  aria-label={t("chat.queueRemove")}
-                  onClick={() => onRemove(message.id)}
-                >
-                  <X className="size-3.5" />
-                </Button>
-              </div>
-            )
-          })}
-        </div>
-      </CollapsibleContent>
-    </Collapsible>
-  )
 }
 
 function paletteLabels({
